Added casting on saome parameters in at25f512b. Added comments and pwm channels number in pwm.h

+++ platform/colibri/dev/at25f512b.c
@@ -7,7 +7,7 @@ void at52f512_read(uint32_t address, uint8_t *buff, uint8_t len){
 	uint8_t busy;
 	AT25F512_WT_BUSY(busy);
 
-	AT25F512_SPI_START(at25f512b_ReadPage);
+	AT25F512_SPI_START((uint8_t)at25f512b_ReadPage);
 	AT25F512_SPI_BYTE_WRITE((address>>16)&0xff);
 	AT25F512_SPI_BYTE_WRITE((address>>8 )&0xff);
 	AT25F512_SPI_BYTE_WRITE((address    )&0xff);
@@ -24,9 +24,9 @@ void at52f512_write(uint32_t address, uint8_t *buff, uint8_t len){
 	uint8_t busy;
 	AT25F512_WT_BUSY(busy);
 
-	AT25F512_SPI_BYTE_CMD(at25f512b_Wren);
+	AT25F512_SPI_BYTE_CMD((uint8_t)at25f512b_Wren);
 
-	AT25F512_SPI_START(at25f512b_WritePage);
+	AT25F512_SPI_START((uint8_t)at25f512b_WritePage);
 	AT25F512_SPI_BYTE_WRITE((address>>16)&0xff);
 	AT25F512_SPI_BYTE_WRITE((address>>8 )&0xff);
 	AT25F512_SPI_BYTE_WRITE((address    )&0xff);
@@ -45,10 +45,10 @@ void at25f512_erase(at25f512b_instr blockSize, uint32_t address){
 	uint8_t busy;
 	AT25F512_WT_BUSY(busy);
 
-	AT25F512_SPI_BYTE_CMD(at25f512b_Wren);
+	AT25F512_SPI_BYTE_CMD((uint8_t)at25f512b_Wren);
 
-	if (blockSize == at25f512b_EraseBulk){
-		AT25F512_SPI_BYTE_CMD(at25f512b_EraseBulk);
+	if (blockSize == (uint8_t)at25f512b_EraseBulk){
+		AT25F512_SPI_BYTE_CMD((uint8_t)at25f512b_EraseBulk);
 	}
 	else{
 		AT25F512_SPI_START(blockSize);
@@ -64,9 +64,9 @@ void at25f512_wrOTP(uint32_t address, uint8_t *buff, uint8_t len){
 	uint8_t busy;
 	AT25F512_WT_BUSY(busy);
 
-	AT25F512_SPI_BYTE_CMD(at25f512b_Wren);
+	AT25F512_SPI_BYTE_CMD((uint8_t)at25f512b_Wren);
 
-	AT25F512_SPI_START(at25f512b_WrOtp);
+	AT25F512_SPI_START((uint8_t)at25f512b_WrOtp);
 	AT25F512_SPI_BYTE_WRITE((address>>16)&0xff);
 	AT25F512_SPI_BYTE_WRITE((address>>8 )&0xff);
 	AT25F512_SPI_BYTE_WRITE((address    )&0xff);
@@ -84,7 +84,7 @@ void at25f512_rdOTP(uint32_t address, uint8_t *buff, uint8_t len){
 	uint8_t busy;
 	AT25F512_WT_BUSY(busy);
 
-	AT25F512_SPI_START(at25f512b_ReadOtp);
+	AT25F512_SPI_START((uint8_t)at25f512b_ReadOtp);
 	AT25F512_SPI_BYTE_WRITE((address>>16)&0xff);
 	AT25F512_SPI_BYTE_WRITE((address>>8 )&0xff);
 	AT25F512_SPI_BYTE_WRITE((address    )&0xff);
@@ -102,7 +102,7 @@ void at25f512_rdOTP(uint32_t address, uint8_t *buff, uint8_t len){
 
 uint8_t at25f512_Rdsr(){
 	uint8_t sr;
-	AT25F512_SPI_START(at25f512b_Rdsr);
+	AT25F512_SPI_START((uint8_t)at25f512b_Rdsr);
 	AT25F512_SPI_BYTE_READ(&sr);
 	AT25F512B_SPI_DISABLE();
 	return sr;
@@ -115,7 +115,7 @@ uint16_t at25f512_RdID(){
 	uint8_t busy;
 	AT25F512_WT_BUSY(busy);
 
-	AT25F512_SPI_START(at25f512b_RdID);
+	AT25F512_SPI_START((uint8_t)at25f512b_RdID);
 	for(i=0;i<2;i++)
 		AT25F512_SPI_BYTE_READ((((uint8_t*)&id)+i));
 	AT25F512B_SPI_DISABLE();
@@ -125,7 +125,7 @@ uint16_t at25f512_RdID(){
 void at25f512_DP(at25f512b_dpMode mode){
 	//FIXME: aggiungere un wait??
 	if (mode == at25f512b_DPon)
-		AT25F512_SPI_BYTE_CMD(at25f512b_DP); //entering dp:3us
+		AT25F512_SPI_BYTE_CMD((uint8_t)at25f512b_DP); //entering dp:3us
 	else
-		AT25F512_SPI_BYTE_CMD(at25f512b_ResumeDP); //exiting dp: 8us
+		AT25F512_SPI_BYTE_CMD((uint8_t)at25f512b_ResumeDP); //exiting dp: 8us
 }
